+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added Tenant Name variable being able to have Spark Dispatcher with different name
 * Enable Debug mode
+* Support Dynamic allocation with Calico (External shuffle working in HOST mode)
 
 ## 2.2.0.3 (December 27, 2017)